refactor(flows): move piece upgrade endpoint into its own module, expose on ee
